// src/shared/api/generated/__swagger_client/models/ItemCreateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemCreateRequest = {
    code: string;
    name: string;
    groupId?: string;
    baseUnitId: string;
    barcode?: string;
    weightKg?: number;
    volumeM3?: number;
    active?: boolean;
};